independent citation generator (#835)
