+++ qutip/about.py
@@ -33,7 +33,7 @@
 """
 Command line output of information on QuTiP and dependencies.
 """
-__all__ = ['about', 'cite']
+__all__ = ['about']
 
 import sys
 import os
@@ -82,55 +82,16 @@ def about():
                                            platform.machine()))
     qutip_install_path = os.path.dirname(inspect.getsourcefile(qutip))
     print("Installation path:  %s" % qutip_install_path)
+
+    # citation
+    longbar = "=============================================================="
+    longbar += "================"
+    cite_msg = "For your convenience a bibtex file can be easily generated"
+    cite_msg += " using `qutip.cite()`"
+    print(longbar)
     print("Please cite QuTiP in your publication.")
-    print("For your convenience a bibtex file can be easily generated using `qutip.about.cite()`")
-
-def cite(path=None, verbose=True):
-    """
-    Citation information and bibtex generator for QuTiP
-
-    Parameters
-    ----------
-    path: str
-        The complete directory path to generate the bibtex file.
-        If not specified then the citation will be generated in cwd
-    """
-    citation = ["@article{qutip2,",
-                "doi = {10.1016/j.cpc.2012.11.019},",
-                "url = {https://doi.org/10.1016/j.cpc.2012.11.019},",
-                "year  = {2013},",
-                "month = {apr},",
-                "publisher = {Elsevier {BV}},",
-                "volume = {184},",
-                "number = {4},",
-                "pages = {1234--1240},",
-                "author = {J.R. Johansson and P.D. Nation and Franco Nori},",
-                "title = {{QuTiP} 2: A Python framework for the dynamics of open quantum systems},",
-                "journal = {Computer Physics Communications}",
-                "}",
-                "@article{qutip1,",
-                "doi = {10.1016/j.cpc.2012.02.021},",
-                "url = {https://doi.org/10.1016/j.cpc.2012.02.021},",
-                "year  = {2012},",
-                "month = {aug},",
-                "publisher = {Elsevier {BV}},",
-                "volume = {183},",
-                "number = {8},",
-                "pages = {1760--1772},",
-                "author = {J.R. Johansson and P.D. Nation and Franco Nori},",
-                "title = {{QuTiP}: An open-source Python framework for the dynamics of open quantum systems},",
-                "journal = {Computer Physics Communications}",
-                "}"]
-
-    if verbose:
-        print("\n".join(citation))
-
-    if not path:
-        path = os.getcwd()
-
-    filename = "qutip.bib"
-    with open(os.path.join(path, filename), 'w') as f:
-        f.write("\n".join(citation))
+    print(longbar)
+    print(cite_msg)
 
 
 if __name__ == "__main__":
